Safari-609.1.11.1/LayoutTests/js/eval-json-proto-expected.txt

This test case checks different JSON literal constructions and ensures they handle __proto__ as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ({__proto__:[]}) instanceof Array is true
PASS evalResult instanceof Array is true
PASS jsonParseResult instanceof Array is false
PASS successfullyParsed is true

TEST COMPLETE
